Cleans up code to 0 warnings, adds image files
Adds the animated status images supposed to be in a few commits back
Moves trigger mask to separate header file (no reason for it to be in the protocol itself)
Cleans up the headers for classes. Prototypes reflect what the object actually uses.
Fixes some random jackConnected() in AppController. Not sure why it was like that...